requirements.txt
supabase==2.3.0
python-dotenv==1.0.0
postgrest==0.13.0
realtime==1.0.0
storage3==0.7.0